use int32

+++ src/meta_preproc.h
@@ -167,12 +167,12 @@ typedef struct ParsedTdfa {
 } ParsedTdfa;
 
 typedef struct ExtractedRegex {
-    int64 source_start_offset;
-    int64 source_end_offset;
+    int32 source_start_offset;
+    int32 source_end_offset;
 
     bool is_null_macro;
 
-    int64 quote_start_offset;
+    int32 quote_start_offset;
     int32 original_string_length;
 
     ParsedOp temp_ops[PREPROC_MAX_TEMP_OPS];
